Simplify Boolean conditional tests
Assuming x is 0 or 1 (false or true), the following are equivalent:
if (x == true)      if (x)
if (x == false)     if (!x)

Simplify the code to avoid explicit tests that the compiler does
internally.

Note that if x is any other integer, it's not ==true, but is true.
However, this fact shouldn't affect any of these tests.