ping_sweep: fix target MAC addresses
Currently, the ICMP pings are sent to a MAC address of zero, so the
recipient won't see the packets; their NIC will drop them.

When performing a network-wide ping sweep, use an ARP sweep rather than
an ICMP sweep. We must ARP each address before performing an ICMP ping
anyway, or we don't know the MAC address to send the ICMP to. If the ARP
succeeds, the sniffer has already seen that packet, so there's no point
sending an ICMP ping afterwards. If the ARP fails, we can't send an ICMP
ping since we don't know where to send it to.

Once we know the target's IP address and are sending periodic ICMP pings,
force the destination MAC address so the packets get sent to the correct
place.