sniffer: protect against re-entrant calls
When we get rid of customPacketLoop() in the next commit, the sniffer
may be called re-entrantly during sniffer_callback()'s call to
turn_my_house_on(). This can use unbounded amounts of stack. Since we
know that we're turning on the house anyway, we don't care about sniffing
any additional packets during the nested calls, so ignore any nested
calls.