Explicity make SECONDS long
Without this, all the literal values in the time calculations are small,
and hence not long. The math operations performed on them (e.g.
1 * SECONDS) don't cause any type promotion to larger types. So, many of
the calculations overflow their types and cause the actual timeouts to be
extremely short. Explicitly forcing seconds to be a long ensures that all
of the calculations are performed within a long and do not overflow, thus
yielding the desired values.

When the timeouts were written in full as a literal before, I believe the
values were large enough to force the types to be a long, so this problem
didn't occur.

This feels like a bug in the compiler (if the default type for literals is
16-bit, then any literal that won't fit into 16-bit and isn't explicitly
marked as a long should either be truncated too or cause a warning/eror).
However, I'm not enough of a C language lawyer to say for sure. Plus I'm
used to 32-bit systems where I wouldn't notice such issues.