Store target_ip in the EEPROM
When the system reboots, it looses track of target_ip, since it was in
RAM. Store the target_ip in EEPROM to prevent this.

To ensure a graceful upgrade path, add a checksum to the EEPROM data.
Otherwise, the first time the code was run after the upgrade, it would
pick up "random" data as the IP address.

The only disadantage here is that if the etherhouse is left off for a
long time, the IP might become stale, yet the etherhouse would still
remember it, and be ICMP pinging the target for ABSENSE_TIMEOUT before
going back to ARP sweeps. However, I suspect this is a less common case.
If the device actually is on the network, it'll soon transmit some
broadcast packet anyway, and target_ip will fall back into sync.